Struct querystrong::QueryStrong

pub struct QueryStrong(_);

Implementations

impl QueryStrong

pub fn new() -> Self

Creates a new (empty) querystrong that contains a map as the top level value

pub fn parse(s: &str) -> Result<Self>

Attempts to create a build a querystrong from the supplied querystring